Refactored serializer utils into separate files
