+++ firstvoices/backend/serializers/site_serializers.py
@@ -11,7 +11,7 @@ from backend.serializers.base_serializers import (
 )
 from backend.serializers.fields import SiteViewLinkField
 from backend.serializers.media_serializers import ImageSerializer, VideoSerializer
-from backend.serializers.utils import get_site_from_context
+from backend.serializers.utils.context import get_site_from_context
 from backend.serializers.validators import SameSite
 from backend.serializers.widget_serializers import SiteWidgetListSerializer
